Made findEventsByDate() and getEvent() functional.
findEventsByDate() now starts API request on the date that matches the starting date results page. The Meetup API is repeatedly called to fetch consecutive pages of results until there are no more or the end date has been matched. A separate call to getEvent() is loaded via ajax to populate a google infoWindow at the time of opening. This replaces loading event descriptions for all events in a single page load.